move entry point file back
