+++ src/mina-signer/tests/stake-delegation.test.ts
@@ -1,4 +1,4 @@
-import Client from '../dist/mina-signer/src/MinaSigner.js';
+import Client from '../dist/mina-signer/MinaSigner.js';
 import type {
   Keypair,
   Signed,
