+++ tsconfig.mina-signer.json
@@ -1,5 +1,5 @@
 {
-  "include": ["./src/mina-signer/src/MinaSigner.ts"],
+  "include": ["./src/mina-signer/MinaSigner.ts"],
   "exclude": ["./src/**/*.unit-test.ts"],
   "compilerOptions": {
     "rootDir": "./src",
